created general apiPost.ts method file
