+++ Frontend/todo-front-end/src/api/apiPost.ts
@@ -1,0 +1,17 @@
+
+
+
+export default function apiPost(endpoint: string, dto : any) {
+    const requestHeaders: HeadersInit = new Headers({
+        'Content-Type': 'application/json'
+    });
+
+    let url: string = `${process.env.REACT_APP_API_URL}${endpoint}`;
+    const requestOptions: RequestInit = {
+        method: 'POST',
+        headers: requestHeaders,
+        body: JSON.stringify(dto)
+    };
+
+    return fetch(url, requestOptions);
+}
