+++ Frontend/todo-front-end/src/api/authentication.ts
@@ -1,4 +1,5 @@
 import React from "react";
+import apiPost from "./apiPost";
 
 const registerEndpoint: string = '/authentication/register';
 
@@ -22,27 +23,16 @@ export const useApiRegister
     = (setStatus: React.Dispatch<RegistrationStatus>,
        onSuccess : (response: RegisterRequestResponse) => void) =>
     (dto: RegisterDto) => {
-        const requestHeaders: HeadersInit = new Headers({
-            'Content-Type': 'application/json'
-        });
-
-        const requestOptions: RequestInit = {
-            method: 'POST',
-            headers: requestHeaders,
-            body: JSON.stringify(dto)
-        };
-
-        let url: string = `${process.env.REACT_APP_API_URL}${registerEndpoint}`;
-        fetch(url, requestOptions)
+        apiPost(registerEndpoint, dto)
             .then(response => {
                 if(!response.ok)
                     throw new Error(response.statusText);
 
-                return response.json()
+                return response.json();
             })
             .then(responseJson => {
                 let response = responseJson as RegisterRequestResponse;
-                onSuccess(response)
+                onSuccess(response);
                 setStatus({loading: false, error: false});
             })
             .catch(() => {
